# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fincava — Colombian Agricultural B2B Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            type: string
        - name: certification
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minOrderKg
          in: query
          schema:
            type: number
        - name: minCupping
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{id}/similar:
    get:
      operationId: getSimilarProducts
      tags: [products]
      summary: Get similar products
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Similar products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /supplier/products:
    get:
      operationId: listMyProducts
      tags: [products]
      summary: List supplier's own products
      responses:
        "200":
          description: Supplier's products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (supplier only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /supplier/products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (supplier only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (supplier only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # SUPPLIERS
...
  /products/{id}/reviews:
    get:
      operationId: listProductReviews
      tags: [reviews]
      summary: List reviews for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # MESSAGES
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        supplierName:
          type: string
        supplierVerified:
          type: boolean
        supplierLogoUrl:
          type: ["string", "null"]
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        description:
          type: string
        origin:
          type: string
        altitude:
          type: ["string", "null"]
        process:
          type: ["string", "null"]
        variety:
          type: ["string", "null"]
        minOrderKg:
          type: number
        maxOrderKg:
          type: ["number", "null"]
        pricePerKgUSD:
          type: number
        availableKg:
          type: number
        harvestSeason:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        cupping:
          type: ["number", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - companyId
        - supplierName
        - supplierVerified
        - name
        - category
        - description
        - origin
        - minOrderKg
        - pricePerKgUSD
        - availableKg
        - images
        - certifications
        - active
        - featured
        - reviewCount
        - createdAt
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            supplierDescription:
              type: ["string", "null"]
            supplierCountry:
              type: ["string", "null"]
            supplierRegion:
              type: ["string", "null"]
            supplierWebsite:
              type: ["string", "null"]
            supplierMemberSince:
              type: ["string", "null"]
            supplierCertifications:
              type: array
              items:
                $ref: "#/components/schemas/Certification"
            originStory:
              type: ["string", "null"]
            farmerName:
              type: ["string", "null"]
          required:
            - supplierCertifications
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - products
        - total
        - page
        - limit
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        description:
          type: string
        origin:
          type: string
        altitude:
          type: ["string", "null"]
        process:
          type: ["string", "null"]
        variety:
          type: ["string", "null"]
        minOrderKg:
          type: number
        maxOrderKg:
          type: ["number", "null"]
        pricePerKgUSD:
          type: number
        availableKg:
          type: number
        harvestSeason:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        cupping:
          type: ["number", "null"]
        originStory:
          type: ["string", "null"]
        farmerName:
          type: ["string", "null"]
      required:
        - name
        - category
        - description
        - origin
        - minOrderKg
        - pricePerKgUSD
        - availableKg
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        subCategory:
          type: ["string", "null"]
        description:
          type: string
        origin:
          type: string
        altitude:
          type: ["string", "null"]
        process:
          type: ["string", "null"]
        variety:
          type: ["string", "null"]
        minOrderKg:
          type: number
        maxOrderKg:
          type: ["number", "null"]
        pricePerKgUSD:
          type: number
        availableKg:
          type: number
        harvestSeason:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        cupping:
          type: ["number", "null"]
        active:
          type: boolean
        featured:
          type: boolean
        originStory:
          type: ["string", "null"]
        farmerName:
          type: ["string", "null"]
    Certification:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        issuer:
          type: string
        expiryDate:
          type: ["string", "null"]
        documentUrl:
          type: ["string", "null"]
        verified:
          type: boolean
      required:
        - id
        - type
        - issuer
        - verified
...
    Review:
      type: object
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        authorCountry:
          type: ["string", "null"]
        productId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        verified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - authorId
        - authorName
        - productId
        - rating
        - verified
        - createdAt
    CreateReviewBody:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required:
        - rating
...