# lib/api-zod/src/generated/types/createProductBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fincava — Colombian Agricultural B2B Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProductBody {
  name: string;
  category: string;
  /** @nullable */
  subCategory?: string | null;
  description: string;
  origin: string;
  /** @nullable */
  altitude?: string | null;
  /** @nullable */
  process?: string | null;
  /** @nullable */
  variety?: string | null;
  minOrderKg: number;
  /** @nullable */
  maxOrderKg?: number | null;
  pricePerKgUSD: number;
  availableKg: number;
  /** @nullable */
  harvestSeason?: string | null;
  images?: string[];
  certifications?: string[];
  /** @nullable */
  cupping?: number | null;
  /** @nullable */
  originStory?: string | null;
  /** @nullable */
  farmerName?: string | null;
}