# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fincava — Colombian Agricultural B2B Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  companyId: number;
  supplierName: string;
  supplierVerified: boolean;
  /** @nullable */
  supplierLogoUrl?: string | null;
  name: string;
  category: string;
  /** @nullable */
  subCategory?: string | null;
  description: string;
  origin: string;
  /** @nullable */
  altitude?: string | null;
  /** @nullable */
  process?: string | null;
  /** @nullable */
  variety?: string | null;
  minOrderKg: number;
  /** @nullable */
  maxOrderKg?: number | null;
  pricePerKgUSD: number;
  availableKg: number;
  /** @nullable */
  harvestSeason?: string | null;
  images: string[];
  certifications: string[];
  /** @nullable */
  cupping?: number | null;
  active: boolean;
  featured: boolean;
  /** @nullable */
  avgRating?: number | null;
  reviewCount: number;
  createdAt: string;
}